heroku/requirements.txt
line-bot-sdk==1.5.0
future==0.16.0
Flask== 0.12.2
click==6.7
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==0.23
Werkzeug==0.12.2
beautifulsoup4==4.7.1
requests==2.20.1
bottle==0.12.15
